[vim] move plugins init fn...
it appeared as if the vimrc bombed out when the colourscheme (installed
by plug) wasn't found, and the plugin fn never got run. Hoping that
moving this before the colourscheme declaration will get initial
plugins even if we have some other problem where the colourscheme isn't
working.

The downside here is having initialisation code in the mieven if we have
some other problem where the colourscheme isn't working.

The downside here is having initialisation code in the middle of the rc
file, i guess we could move the fn and make a fn call...