[brew] deal with xcode inital requirement debacle...
is the requirement for full xcode just for macvim build?

here we try to loudly ensure you have some kind of xcode before we'll
run, and we quietly try and accept its license either way (as its a
silent  no op if license is already accepted)

A requirement for the license was sudo, which actually the rest of the
script was crying out for since we randomly get asked for sudo about 4
times during the installations. Copied the nice sudo persister from the
macos script