[vim] NERDTree open in prev window regardless...
of whether the buffer is already open somewhere else ie: don't switch to
the tab of the previous buffer-instances window and distract me from
what i'm doing. (vim is mean to be able to open multiple views onto the
same buffer). Note needed to do this to LMB click also